Sveva/profile page (#91)
# Profile Page Update

## Overview
I created a profile page that is used for both personal and public
visibility. None of the buttons work; they are just for display. I
started but was overcomplicating it and got lost. I added events for all
buttons, but I am not doing anything.

## Navigation
When clicking on the profile (orange) avatar in the upper right, it
always gives you a page with three buttons: **register**, **login**, and
**profile**. The profile is temporary, so I didn't bother to change the
style. I thought it would be easier, so I don't have to refresh the page
all the time.

## Future Implementation
I will implement the logic to display either register/login options or
take you to your profile if logged in afterward (I think that will be
another issue; need authentication, etc.).

## Current Profile Page Settings
Right now, the profile page is set up as if you are logged in, so
personal visibility means you see settings, can perform actions related
to friends (on the friend page), and you also see your username, avatar
is orange. For others' visibility, you shouldn't be able to see
settings, have no username displayed, and lack access to friend-related
actions on the friend page, avatar is blue. I left the button in the
header, thinking it can serve as another way to perform actions. (Am I
clear?)

## Rendering Logic
Since I will need to render based on the server's response, I haven't
implemented this logic yet (I am working on it). To see the different
pages, you need to go to `accountDisplay.ts` (from line 99) and change
all the style values (I put comments; basically, what is set to flex
goes to none, and what is set to none goes to flex in order to see the
other page, and change colour for avatar).

## Guidelines
I followed the rules on **issue #84**, and I think I have addressed all
of them; I hope I didn't forget anything.

## Style Adjustments
Additionally, the style can be changed and it will be, but I am losing
so much time on it, and I want to be practical first; it's something I
can do later. I added a folder in **images** called "profile," where I
keep the sample avatars, and I hope with different colors, it is easier
to follow.

**Note**: I might have misspelled "display" somewhere; I think I fixed
all instances. It’s been a hard lesson for me...