Sveva/settings profile (#93)
I delete part of html because there was a duplicate: the account div.
Not sure why)
I added the error box at the bottom for registration form and login form
(just to visualise it, it would display only if errors occur)
I change the fields' name in edit information as in #84 and slack.

Two questions: 
1)in #84 username was only readonly, but I think we want it editable
now, right? - super easy fix
2) I would not put logout button in the edit information settings. I
will create a modal page (like in intra when you press on your username)
and I will display it there. I will make another pull request for it.