chat/users tabs (#97)
I worked on the left column of the "chat page."
I added the tabs: one for chats and one for users. By default, it will
display the chats. Clicking on the tab will switch the displayed
content; the tab has a darker color to highlight which content is. Both
chats and users shows the avatar and display name. More features can be
added; I do not remember what we decided.
The right side is not added yet;  that will be in the next PR.