Display conversation or profile in chat page (#98)
I finally finished to implement a temporary logic to display a
conversation or user profile in the chat page.
I used same logic for the chat conversations, for the profile, I am
using one div html element, and I change visibility depending from where
the profile page is requested. I believe I will need to write this page
entirely in js to be able to fill it in with correct info. So, if you go
back to the chat page, you have to click again on the user Eos to see
the profile page (I am just moving around the div element to speed up
the process to see how it would look like).
Also, I noticed the button changed layout in the profile page displayed
in the user chat page. They do it because of less space. So it can be
changed.
I didn't yet implement if clicking on the avatar in the conversation
(upper right corner), don't know if I should display it instead of the
conversation or entire page (most likely the first option).