Create basic flow for game page (#107)
We have a page for the pong game, (gameLogic.ts) when clicking on the
icon in the left navbar.
Right now we can set up a single match between guests. There is the
option for one player to play vs AI, but the AI is not implemented, but
the paddle doesn't move.

I also added the option to write display name/alias. For now, only if
there is some input, then the user is capable to more forward, we will
need to check later.
If we have alias or aliases, then the instructions will be displayed
bellow (to style, but to give an idea). I set a timer to display the
start button game after 5 seconds. I wanted to try it out. Like faking
to prepare the game)
Who scores first 3, wins, then a game over screen appears declaring the
winner, and two buttons (play again or new game). I have not implemented
the buttons in the instructions nor the play again and new game.

I tried not to break the logic of the other pages hide/show based on
interaction, everything seems to work

About the game, I only check for collisions with the paddle, not yet
precise if it hits edges, change direction of the ball, and also it just
start from the centre every time it touches a wall. The sizes, the speed
the colors can be changed. don't know if we should create a option page
(but maybe doing too much), so at least can decide how many points to
play and difficulty (ball speed).
The canva size, the black board to be clear, must be declared, and I
wouldn't make it bigger: then I get very poor pixel resolution, and I
still dont get how to fix that (lost 1 day for it, with trying to create
a modal page...)

Notes:
- for the game I am using setInterval, but I will change it with
requestAnimationFrame();
- need to implement buttons in instructions and in the game over screen
- reset properly the game if it is exited.
- decide if we want to display the nicknames on the board while playing
- make ball bounce more accurate and change direction based on that

Extra
I changed the icons with more readable ones. I am aware they are three
different sizes, just to see which one fits better.
Tournament page isn't implemented, so nothing happens if you try to
click.