Add tournament logic (#126)
It is possible to play one single games (2 players) and to play a
tournament with 4 players (3 matches in total) only for guests.
Each game session is valid only if the game or tournament is finished,
if not, like switching to a different page, I just reset everything. A
warning message can be added.

How it works now:
1) get the game type:  one game or tournament (icon on nav bar)
2) display page with 2 players (one game) or 4 players(tournament) to
write alias or select AI.
3) check alias (right now only duplicates within the game session)
4) add players into a players' list, so I then can create a game based
on the players' info.
4) start the game only if all players have provided an alias
5) show page with rules: first to score to 3 (it can be changed) wins;
who is on which side and they keys to use to move up and down
6) start game button : displays game and play
7) game over screen: displays the winner (it can be improved)
8) if it is a tournament, a new page will display the next match (score,
players and sides)
9) if single match or tournament ends, it just display a button to start
a new game

A bit about the tournament. Right now the matching selects following who
entered their alias first: example, if the order is Eos - Felix - Bianca
- lucky, the pairs will be Eos - felix, Bianca - lucky. The hard part
was to correctly draw the game for each game; The function that draws is
asynchronous, so it is not sequential, and I was not getting it right. I
add each game to a game list, so it is easier to call. I will need to
rewrite this I think since I will need to send it over to the backend
for storage.

For testing, I just start a game, play till the end and press the button
New Game; I also switch pages while playing to make sure nothing is
extra on the page. Everything works as expected, but I could have missed
something, like not properly resetting before/after a game or players'
name are wrong.

What I will do next about tournament
- better/smoother game features (ball collision more precise)
- implement structure to hold player/game/tournament information needed
to be stored for registered users.
- add logic to send request to backend to store
- enhance pages within the game (game preview, game over screens).