Check names for AI and users (#133)
If the AI option is ON, it is not possible to modify the input anymore,
and it write automatically the name of the AI_ followed by Alena, Sveva
or Luis. If press OFF, then you can enter name of your choice.
The box becomes red if the name is invalid and green if valid.
The user can't call him/herself AI or have any ai in their name, but
maybe too strict, it can be adjusted.
Logic to be unique name should still be valid