New Profile Page (#163)
I reorganise the profile page.
Now the Profile, accessible from the dropdown menu, will display the
avatar, username, display name on the top, and on the bottom directly
the stats: 2 graphs and the list of matches played. That's it. I removed
the online status. Maybe I will then remove the third column, but I will
leave it for now.
You can find the Settings in the dropdown menu. I fixed the form: now
each edit will have its own separate request. The username is readOnly,
but I have't filled in. If you try to type, nothing happens. You can
change the avatar by clicking on it. it will show a popup: either
default one, or the + for upload. To close, press the big X on the left.
You can find Friends in the dropdown menu. I will maybe rearrange them,
but I will check after being able to retrieve them from server. Don't
have it in me to hardcode 10 friends.
All forms and buttons are disable, to prevent breaking code. It will be
fixed in the next PR, when connecting to server.
Background colors can be changed with different colros or images.
Everything should work as before: change page and not displaying
elements from the previous page. Once the login/logout logic is
implemented, the dropdown menu will look smaller)) now all pages are
there for easier access.
Quick note: I do not remember what we agreed for the stats, and the
issue was not helpful. They can be changed too #152