Login user (#171)
Login form works: it logins in the user, change the header with the name
and avatar picture (that is hardcode it at the moment since it comes as
null).
It displays errors if something goes wrong with username or password. 
The dropdown menu should display register / login if none is logged in,
and show profile, settings, friends and logout if logged in.