User profile (#172)
To be able to request after logging in and use the cookie, I had to
change a line in main.ts
`fastify.register(require('@fastify/cors'), { origin: '*' })
//https://github.com/fastify/fastify-cors`
in
`fastify.register(require('@fastify/cors'), { origin:
['http://127.0.0.1:8081', 'http://localhost:8081'], credentials: true})
`
(or write your domain and port, my 8080 is taken by I dont know who..)
or I get this error in the console
`Cross-Origin Request Blocked: The Same Origin Policy disallows reading
the remote resource at ‘http://127.0.0.1:3000/users/me’. (Reason:
Credential is not supported if the CORS header
‘Access-Control-Allow-Origin’ is ‘*’)`
I need to pass the credentials header in the request to be able to
retrieve and send the cookie you send me after login

The user can request 
1)personal profile -> update of username and display name - stats not
linked yet
2)settings page -> user can change display name, password and delete
account and upload image
Problems for images:
- uploading works, but I am not able to display the avatar because the
avatarUrl is a path to the image and not url - so you will get null in
the console when trying to get the image, it wont display.
- how do we handle default pictures? they dont need to be uploaded since
they already exists. So, either I download the default picture, then
uploaded it again to the user folder, so it follows the same procedure
than the upload image, or we provide the url to the default picture. By
doing so, I can't make same post request since I only have a url and no
data
3)friends page -> display list of friends, buttons remove, block friend
correctly, also display the blocked friends.
Problem
- if a friends is being blocked, it is still in the friend list, so when
getting the friends list, it will be displayed in both blocked and
friend list.

Other problems:
- sometimes I get error message for` settingsUsernameInput.value`, I was
not able to determine when and why, if it happens let me know. Sometimes
works other doesnt...

To test the friend's buttons, need to add with curl first...missing that
button since I haven't done user page yet