AI Implementation (#132)
We do have an AI opponent! They can also play against each other.
It follows the ball y, and it moves up or down depending if the ball y
is above or below the middle of the paddle. The AI paddle only moves
when the ball is in its side (it can be changed, but that is how I
play)).
I haven't set different level of AI opponent yet, I feel sometimes is
too easy, but other times harder, not sure why.
I also fixed the problem with the ball colliding with the edges of the
paddle. So it keeps playing.
I would like to make only last implementation to the ball, to make it
less predictable